+++ fw_beam/pid_control.h
@@ -1,6 +1,8 @@
 #ifndef PID_CONTROL_H
 #define PID_CONTROL_H
 
+#include <stdint.h>
+
 class PIDController {
 private:
     double kp;
@@ -15,7 +17,7 @@ private:
     double last_input;
 public:
     PIDController(double kp, double ki, double kd, double dt, double target, double out_min, double out_max);
-    bool calculate(double *input, double *response);
+    bool calculate(uint16_t *input, double *response);
 };
 
 #endif // PID_CONTROL_H
